+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## unreleased
 
 - Casing - Added support for casing of the required properties (#77)
+- Bumped Jest dependency from 27.5.1 to 29.7.0
 
 ## [1.14.0] - 2023-04-11